Extract GP training helpers module
