+++ main_eier.py
@@ -7,15 +7,19 @@ import time
 
 import numpy as np
 import yaml
-from sklearn.gaussian_process import GaussianProcessRegressor
-from sklearn.gaussian_process.kernels import Matern, ConstantKernel
 from scipy.stats import norm
 import wandb
 
 from limit_states import REGISTRY as ls_REGISTRY
 from active_learning.active_learning import AcquisitionStrategy
 from active_learning.eier import build_gp_cache_from_gpr, estimate_pf_posterior_samples
-from utils.data import isoprobabilistic_transform, custom_optimizer, normalize_array, parallel_predict
+from utils.data import isoprobabilistic_transform, normalize_array, parallel_predict
+from utils.gp_training import (
+    fit_gp_with_optional_stabilization,
+    is_bad_fit,
+    make_base_kernel,
+    resolve_gp_alpha,
+)
 
 
 def _fmt_sci(value):
@@ -27,31 +31,6 @@ def _fmt_sci(value):
     if np.isneginf(value):
         return "-inf"
     return f"{value:.3E}"
-
-
-def make_base_kernel(input_dim, upper_bound=1e5):
-    upper_bound = float(upper_bound)
-    length_init = np.full(input_dim, 1.0, dtype=np.float64)
-    kernel = ConstantKernel(1.0, (1e-5, upper_bound)) * Matern(
-        length_scale=length_init,
-        length_scale_bounds=(1e-5, upper_bound),
-        nu=2.5,
-    )
-    return kernel
-
-
-def is_bad_fit(current_lml, prev_lml, lml_drop_tol=50.0, abs_lml_low=-100.0):
-    """
-    Consider a fit 'bad' if:
-      - LML drops a lot compared to the previous good model, OR
-      - LML is absolutely very low.
-    """
-    too_low = current_lml < abs_lml_low
-    if prev_lml is None:
-        big_drop = False
-    else:
-        big_drop = current_lml < (prev_lml - lml_drop_tol)
-    return too_low or big_drop
 
 
 def _resolve_cpu_workers(value):
@@ -84,68 +63,6 @@ def _resolve_cpu_workers(value):
             "for this task. Capping worker count."
         )
     return int(min(value, available_workers))
-
-
-def _resolve_gp_alpha(config, y_train):
-    y_scale = max(float(np.std(y_train)), 1e-12)
-    alpha = float(config['obs_stddev'] / y_scale) ** 2
-    return max(alpha, float(config.get('min_gp_alpha', 1e-12)))
-
-
-def _fit_gp_with_optional_stabilization(
-    x_train_norm,
-    y_train,
-    init_kernel,
-    input_dim,
-    gp_alpha,
-    n_restarts_optimizer,
-    enable_stabilization=False,
-    max_gp_alpha=1e-4,
-):
-    def _build_model(kernel, alpha):
-        return GaussianProcessRegressor(
-            kernel=kernel,
-            n_restarts_optimizer=n_restarts_optimizer,
-            normalize_y=True,
-            optimizer=custom_optimizer,
-            alpha=float(alpha),
-        )
-
-    model_gp = _build_model(init_kernel, gp_alpha)
-    try:
-        model_gp.fit(x_train_norm, y_train)
-        return model_gp, {
-            "stabilized": False,
-            "alpha_used": float(gp_alpha),
-            "retry_count": 0,
-            "kernel_upper": 1e5,
-        }
-    except np.linalg.LinAlgError:
-        if not enable_stabilization:
-            raise
-
-    retry_count = 0
-    gp_alpha_retry = max(float(gp_alpha), 1e-12)
-    while gp_alpha_retry <= float(max_gp_alpha) + 1e-18:
-        for kernel_upper in (1e5, 30.0):
-            retry_count += 1
-            retry_kernel = make_base_kernel(input_dim, upper_bound=kernel_upper)
-            retry_model = _build_model(retry_kernel, gp_alpha_retry)
-            try:
-                retry_model.fit(x_train_norm, y_train)
-                return retry_model, {
-                    "stabilized": True,
-                    "alpha_used": float(gp_alpha_retry),
-                    "retry_count": int(retry_count),
-                    "kernel_upper": float(kernel_upper),
-                }
-            except np.linalg.LinAlgError:
-                continue
-        gp_alpha_retry *= 10.0
-
-    raise np.linalg.LinAlgError(
-        "GP fit failed after stabilization retries (alpha backoff + tighter kernel bounds)."
-    )
 
 
 def main(config, name_exp):
@@ -305,9 +222,9 @@ def main(config, name_exp):
 
         gp_alpha = 1e-8
         if al_strategy == "eier":
-            gp_alpha = _resolve_gp_alpha(config, y_train)
+            gp_alpha = resolve_gp_alpha(config, y_train)
 
-        model_gp, gp_fit_info = _fit_gp_with_optional_stabilization(
+        model_gp, gp_fit_info = fit_gp_with_optional_stabilization(
             x_train_norm=x_train_norm,
             y_train=y_train,
             init_kernel=init_kernel,
@@ -321,7 +238,7 @@ def main(config, name_exp):
         if is_bad_fit(lml, lml_prev, lml_drop_tol=50.0, abs_lml_low=-100.0):
             # This fit looks suspicious -> try a fresh base kernel with restarts
             base_kernel = make_base_kernel(lstate.input_dim)
-            model_gp_fresh, gp_fit_info_fresh = _fit_gp_with_optional_stabilization(
+            model_gp_fresh, gp_fit_info_fresh = fit_gp_with_optional_stabilization(
                 x_train_norm=x_train_norm,
                 y_train=y_train,
                 init_kernel=base_kernel,
